+++ cli/cli/commands/engine/engine.go
@@ -2,6 +2,7 @@ package engine
 
 import (
 	"github.com/kurtosis-tech/kurtosis-cli/cli/commands/engine/start"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/commands/engine/stop"
 	"github.com/spf13/cobra"
 )
 
@@ -13,4 +14,5 @@ var EngineCmd = &cobra.Command{
 
 func init() {
 	EngineCmd.AddCommand(start.StartCmd)
+	EngineCmd.AddCommand(stop.StopCmd)
 }