+++ cli/docs/changelog.md
@@ -3,9 +3,11 @@
 * Added documentation in README about how to develop on this repo
 * Upgraded to `kurtosis-core` 1.25.2, which contains fixes for `container-engine-lib` 0.7.0 feature that allows binding container ports to specific host machine ports
 * Added `engine start` command to the CLI
+* Added `engine stop` command to the CLI
 
 ### Fixes
 * Clarified the difference between the two types of params in `module exec`
+* `engine start` won't start another container if one is already running
 
 # 0.4.2
 ### Features